+++ examples/data-sources/iosxr_flow_sampler_map/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_flow_sampler_map" "example" {
+  true = "sampler_map1"
+}