+++ examples/data-sources/iosxr_sampler_map/data-source.tf
@@ -1,3 +1,0 @@
-data "iosxr_sampler_map" "example" {
-  sampler_map_name = "sampler_map1"
-}