+++ examples/resources/iosxr_flow_sampler_map/resource.tf
@@ -1,0 +1,5 @@
+resource "iosxr_flow_sampler_map" "example" {
+  true   = "sampler_map1"
+  random = 1
+  out_of = 1
+}