+++ examples/resources/iosxr_sampler_map/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_sampler_map.example "Cisco-IOS-XR-um-flow-cfg:/sampler-maps/sampler-map[sampler-map-name=sampler_map1]"