+++ examples/resources/iosxr_sampler_map/resource.tf
@@ -1,5 +1,0 @@
-resource "iosxr_sampler_map" "example" {
-  sampler_map_name = "sampler_map1"
-  random           = 1
-  out_of           = 1
-}